backend/requirements.txt
fastapi==0.124.2
uvicorn[standard]==0.38.0
python-multipart==0.0.20
pydantic-settings==2.12.0
pillow==12.0.0

httpx==0.28.1
